IMDb Movie App
This is an android application that provides information about movies.
It shows all the information about the movies like Actors, Directors,
Genre, Release date according to the descending order of Release date.
The application implements intents to transfer data from one activity to
another. JSON parsing is used for parsing the data from the IMDB
website. Also, ListView is used to display the list of movies using
Adapter class.